Provide MySQL support for implicit datetime hierarchies
Follows t27duck/active_reporting#10, which provided implicit
datetime hierarchies for Postgres.

One improvement that might be made is with respect to the MySQL
implementation of the Postgres date_trunc function. Currently,
the code creates a function in the database to undertake the
date_trunc functionality. However, it seems to me that a better
approach would be to have an anonymous function perform this
action, and avoid polluting a user's database.