Refactor MySQL date_trunc adaption
Favors use of the MySQL "date_add" and "TIMESTAMPDIFF" functions
to generate truncated date strings, akin to the "date_trunc"
Postgres function. While this does provide performance gains and
removes the need for polluting a database with a specialized database
function, it also is much more limited than the PG date_trunc function.
Noteably, dates prior to 1900 will not work correctly, and the intervals
of decade, century, and millenium will not work at all.